AI_Mouse_Controller/tests/test_mouse_control.py
import unittest
from mouse_control import MouseController

class TestMouseController(unittest.TestCase):
    def test_move_mouse(self):
        mouse_controller = MouseController()
        mouse_controller.move_mouse(100, 100)
        # Add assertions here

    def test_click(self):
        mouse_controller = MouseController()
        mouse_controller.click()
        # Add assertions here

    def test_double_click(self):
        mouse_controller = MouseController()
        mouse_controller.double_click()
        # Add assertions here

    def test_right_click(self):
        mouse_controller = MouseController()
        mouse_controller.right_click()
        # Add assertions here

    def test_scroll(self):
        mouse_controller = MouseController()
        mouse_controller.scroll(3)
        # Add assertions here

    def test_drag(self):
        mouse_controller = MouseController()
        mouse_controller.drag(200, 200)
        # Add assertions here

if __name__ == '__main__':
    unittest.main()
